now sorts output file, parse_json fixed

+++ not_following_back.py
@@ -7,11 +7,12 @@ with open('following.json', 'r') as f:
 with open('followers.json', 'r') as f:
   followers = json.load(f)
 
-#Parses the given JSON file into a set
+#Parses the given JSON file into a set of usernames
 def parse_json(json_file, relationship):
-    parsed = {user['relationship_data'][0]['username'] for user in json_file[relationship]}
-    print(parsed)
-    return parsed
+    if type(json_file) == list:
+      return {user['string_list_data'][0]['value'] for user in json_file}
+    else:
+      return {user['string_list_data'][0]['value'] for user in json_file[relationship]}
   
 #Subtracts your followers from your following, to determine who you are following that isn't a also follower of yours 
 not_following = parse_json(following, 'relationships_following') - parse_json(followers, 'relationships_followers')
@@ -22,18 +23,19 @@ if os.path.exists("not_following_back.txt"):
 
 #Writes the list of users to a text file
 with open('not_following_back.txt', 'w') as f:
-  f.writelines([username + '\n' for username in not_following])
+  f.writelines(sorted([username + '\n' for username in not_following]))
 
 #Writes a file with the differences between the new and old instance of not_following, if an old one exists
-with open('not_following_back.txt', "r") as file1, open('old_not_following_back.txt', "r") as file2, open("differences_included.txt", "w") as output_file:
-    file1_contents = file1.readlines()
-    file2_contents = file2.readlines()
-    for user in file1_contents:
-        if user not in file2_contents:
-            output_file.write(f"In old, not in new: {user}")
-    output_file.write("\n")
-    for user in file2_contents:
-        if user not in file1_contents:
-            output_file.write(f"In new, not in old: {user}")
+if os.path.exists("old_not_following_back.txt"):
+  with open('not_following_back.txt', "r") as file1, open('old_not_following_back.txt', "r") as file2, open("differences_included.txt", "w") as output_file:
+      file1_contents = file1.readlines()
+      file2_contents = file2.readlines()
+      for user in file1_contents:
+          if user not in file2_contents:
+              output_file.write(f"In old, not in new: {user}")
+      output_file.write("\n")
+      for user in file2_contents:
+          if user not in file1_contents:
+              output_file.write(f"In new, not in old: {user}")
 
 print("The list of users has been written to 'not_following_back.txt', differences between most recent comparison in 'differences_included.txt'.")
